chore: reorganise, cleanup benchmarks setup,
