+++ .circleci/config.yml
@@ -678,14 +678,7 @@ jobs:
       - save_cache:
           key: cache-{{ .Revision }}-benchmark-app
           paths:
-            - benchmarks/latest/dist/app.latest.js
-            - benchmarks/latest/dist/app-repeat-ce.latest.js
-            - benchmarks/latest/dist/app-big-template.latest.js
-            - benchmarks/local/dist/app.local.js
-            - benchmarks/local/dist/app-repeat-ce.local.js
-            - benchmarks/local/dist/app-big-template.local.js
-            - benchmarks/app-repeat-keyed-string/dist
-            - benchmarks/app-repeat-keyed-expr/dist
+            - benchmarks/*/dist
 
   tacho_benchmark:
     executor: docker_circleci_bench
