+++ benchmarks/app-repeat-ce/rerender-ce.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://raw.githubusercontent.com/Polymer/tachometer/master/config.schema.json",
-  "root": "../../",
+  "root": "../../../",
   
   "benchmarks": [
     {
@@ -10,16 +10,17 @@
         "addArguments": ["--js-flags=--expose-gc", "--enable-precise-memory-info"]
       },
       "measurement": [
-        { "name": "perf", "mode": "performance", "entryName": "startup-1k-big-template" }
+        { "name": "perf", "mode": "performance", "entryName": "rerender-10k" },
+        { "name": "memory", "mode": "expression", "expression": "jsHeapSize" }
       ],
       "expand": [
         {
-          "name": "startup 1k latest (big template)",
-          "url": "startup1k-latest-big-template.html"
+          "name": "rerender CE npm",
+          "url": "rerender10k-ce-latest.html"
         },
         {
-          "name": "startup 1k local (big template)",
-          "url": "startup1k-local-big-template.html"
+          "name": "rerender CE local",
+          "url": "rerender10k-ce-local.html"
         }
       ]
     }
