+++ benchmarks/app-repeat-ce/rerender10k-ce-latest.html
@@ -1,5 +1,5 @@
 <script type="module">
-import { start } from './local/dist/app.local.js';
+import { start } from './dist/app.latest.js';
 
 const host = document.body.appendChild(document.createElement('div'));
 const au = start(host);
@@ -8,9 +8,8 @@ performance.mark('10k-start');
 au.root.controller.viewModel.newItems(10000);
 
 performance.mark('10k-end');
+performance.measure('rerender-10k', '10k-start', '10k-end');
 
 // measure memory in MBs
 window.jsHeapSize = performance.memory.usedJSHeapSize / 1e6;
-
-performance.measure('rerender-10k', '10k-start', '10k-end');
 </script>
