+++ benchmarks/app-repeat-ce/rerender10k-ce-local.html
@@ -1,11 +1,14 @@
 <script type="module">
-import { start } from './local/dist/app-repeat-ce.local.js';
+import { start } from './dist/app.local.js';
 
 const host = document.body.appendChild(document.createElement('div'));
+const au = start(host);
 performance.mark('10k-start');
-const au = start(host, 10000);
+
+au.root.controller.viewModel.newItems(10000);
+
 performance.mark('10k-end');
-performance.measure('startup-10k', '10k-start', '10k-end');
+performance.measure('rerender-10k', '10k-start', '10k-end');
 
 // measure memory in MBs
 window.jsHeapSize = performance.memory.usedJSHeapSize / 1e6;
