+++ benchmarks/app-repeat-ce/startup10k-ce-latest.html
@@ -1,15 +1,12 @@
 <script type="module">
-import { start } from './dist/startup.latest.js';
+import { start } from './dist/app.latest.js';
 
 const host = document.body.appendChild(document.createElement('div'));
-performance.mark('start');
+performance.mark('10k-start');
 const au = start(host, 10000);
-
-performance.mark('end');
-performance.measure('keyed-expr', 'start', 'end');
+performance.mark('10k-end');
+performance.measure('startup-10k', '10k-start', '10k-end');
 
 // measure memory in MBs
 window.jsHeapSize = performance.memory.usedJSHeapSize / 1e6;
-
-
 </script>
