+++ benchmarks/app-repeat-ce/startup10k-ce-local.html
@@ -1,17 +1,12 @@
 <script type="module">
-import { start } from './latest/dist/app.latest.js';
+import { start } from './dist/app.local.js';
 
 const host = document.body.appendChild(document.createElement('div'));
-const au = start(host);
 performance.mark('10k-start');
-
-au.root.controller.viewModel.newItems(10000);
-
+const au = start(host, 10000);
 performance.mark('10k-end');
+performance.measure('startup-10k', '10k-start', '10k-end');
 
 // measure memory in MBs
 window.jsHeapSize = performance.memory.usedJSHeapSize / 1e6;
-
-performance.measure('rerender-10k', '10k-start', '10k-end');
-
 </script>
