+++ benchmarks/app-repeat-view-big-template/rerender1k-latest-big-template.html
@@ -1,14 +1,17 @@
 <script type="module">
-import { start } from './latest/dist/app-big-template.latest.js';
+import { start } from './dist/app.latest.js';
 
 const host = document.body.appendChild(document.createElement('div'));
+const au = start(host);
 performance.mark('1k-start');
-const au = start(host, 500);
+
+au.root.controller.viewModel.newItems(500);
 
 performance.mark('1k-end');
 
 // measure memory in MBs
 window.jsHeapSize = performance.memory.usedJSHeapSize / 1e6;
 
-performance.measure('startup-1k-big-template', '1k-start', '1k-end');
+performance.measure('rerender1k-big-template', '1k-start', '1k-end');
+
 </script>
