+++ benchmarks/app-repeat-view-big-template/startup1k-latest-big-template.html
@@ -1,5 +1,5 @@
 <script type="module">
-import { start } from './local/dist/app-big-template.local.js';
+import { start } from './dist/app.latest.js';
 
 const host = document.body.appendChild(document.createElement('div'));
 performance.mark('1k-start');
