+++ benchmarks/app-repeat-view-big-template/startup1k-local-big-template.html
@@ -1,16 +1,14 @@
 <script type="module">
-import { start } from './local/dist/app-big-template.local.js';
+import { start } from './dist/app.local.js';
 
 const host = document.body.appendChild(document.createElement('div'));
-const au = start(host);
 performance.mark('1k-start');
-
-au.root.controller.viewModel.newItems(500);
+const au = start(host, 500);
 
 performance.mark('1k-end');
 
 // measure memory in MBs
 window.jsHeapSize = performance.memory.usedJSHeapSize / 1e6;
 
-performance.measure('rerender1k-big-template', '1k-start', '1k-end');
+performance.measure('startup-1k-big-template', '1k-start', '1k-end');
 </script>
