+++ benchmarks/app-repeat-view-keyed-expr/startup-latest.html
@@ -1,12 +1,15 @@
 <script type="module">
-import { start } from './latest/dist/app-repeat-ce.latest.js';
+import { start } from './dist/app.latest.js';
 
 const host = document.body.appendChild(document.createElement('div'));
-performance.mark('10k-start');
+performance.mark('start');
 const au = start(host, 10000);
-performance.mark('10k-end');
-performance.measure('startup-10k', '10k-start', '10k-end');
+
+performance.mark('end');
+performance.measure('keyed-expr', 'start', 'end');
 
 // measure memory in MBs
 window.jsHeapSize = performance.memory.usedJSHeapSize / 1e6;
+
+
 </script>
