+++ benchmarks/app-repeat-view-keyed-string/bench.json
