+++ benchmarks/app-repeat-view-keyed-string/index.js
