+++ benchmarks/app-repeat-view-keyed-string/startup-latest.html
@@ -1,5 +1,5 @@
 <script type="module">
-import { start } from './dist/startup.local.js';
+import { start } from './dist/app.latest.js';
 
 const host = document.body.appendChild(document.createElement('div'));
 performance.mark('start');
