+++ benchmarks/app-repeat-view/index.js
