+++ benchmarks/app-repeat-view/rerender.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://raw.githubusercontent.com/Polymer/tachometer/master/config.schema.json",
-  "root": "../../",
+  "root": "../../../",
   
   "benchmarks": [
     {
@@ -10,16 +10,17 @@
         "addArguments": ["--js-flags=--expose-gc", "--enable-precise-memory-info"]
       },
       "measurement": [
-        { "name": "perf", "mode": "performance", "entryName": "rerender1k-big-template" }
+        { "name": "perf", "mode": "performance", "entryName": "rerender-10k" },
+        { "name": "memory", "mode": "expression", "expression": "jsHeapSize" }
       ],
       "expand": [
         {
-          "name": "rerender 1k latest (big template)",
-          "url": "rerender1k-latest-big-template.html"
+          "name": "rerender npm",
+          "url": "rerender10k-latest.html"
         },
         {
-          "name": "rerender 1k local (big template)",
-          "url": "rerender1k-local-big-template.html"
+          "name": "rerender local",
+          "url": "rerender10k-local.html"
         }
       ]
     }
