+++ benchmarks/app-repeat-view/startup.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://raw.githubusercontent.com/Polymer/tachometer/master/config.schema.json",
-  "root": "../../",
+  "root": "../../../",
   
   "benchmarks": [
     {
@@ -10,17 +10,17 @@
         "addArguments": ["--js-flags=--expose-gc", "--enable-precise-memory-info"]
       },
       "measurement": [
-        { "name": "perf", "mode": "performance", "entryName": "rerender-10k" },
+        { "name": "perf", "mode": "performance", "entryName": "startup-10k" },
         { "name": "memory", "mode": "expression", "expression": "jsHeapSize" }
       ],
       "expand": [
         {
-          "name": "10k ce dev",
-          "url": "rerender10k-ce-latest.html"
+          "name": "startup npm",
+          "url": "startup10k-latest.html"
         },
         {
-          "name": "10k ce local",
-          "url": "rerender10k-ce-local.html"
+          "name": "startup local",
+          "url": "startup10k-local.html"
         }
       ]
     }
