+++ benchmarks/app-repeat-view/startup10k-local.html
