+++ benchmarks/app-repeat-view/update1k-latest.html
@@ -1,5 +1,5 @@
 <script type="module">
-import { start } from './latest/dist/app.latest.js';
+import { start } from './dist/app.latest.js';
 
 const host = document.body.appendChild(document.createElement('div'));
 const au = start(host, 1000);
