+++ benchmarks/app-repeat-view/update1k-local.html
@@ -1,5 +1,5 @@
 <script type="module">
-import { start } from './local/dist/app.local.js';
+import { start } from './dist/app.local.js';
 
 const host = document.body.appendChild(document.createElement('div'));
 const au = start(host, 1000);
